+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 * Refactored `KeyboardControlService` docs.
 * Added `maximumNumberOfPoints` to `PolygonEditOptions` - you can now set maximum number of points when creating polygon using `PolygonEditorService` .
 * Refactored `KeyboardControlService` docs. 
+* Added flyTo options to `CameraService.trackEntity()` + example in the demo.
+
+### Fixes
+* `ac-point-desc` changed to Entity instead of Primitive. For primitive use `ac-point-primitive-desc`.
 
 ## 0.0.49
 ### Fixes